chore: add createdAt, updatedAt to CardType
